l04: a quick fix for the CPU problem
We sleep 10ms if we didn't read any data (or accepted a conection).
(Question: why can't we just sleep on every iteration?)

This introduces another problem: if data comes just after
we go to sleep, we will read it up to 10ms too late, which
might be a problem in a high performance application.

Also we are still waking 100 times a second which is not
very good for battery comsuption on a laptop...

And strace-ing this app is a nightmare.