l04: use select to wait for available data
There is one important thing to remember about select: it will
block until _new_ data _arrives_ for a file descriptor (or data
leaves a buffer for writes). Thus, if 100 bytes arrive, select
will finish, but if we read only 50 bytes and call select again,
it will wait indefinitely for new data to arrive even though
there are 50 bytes waiting in the buffer!

This is the reason why we are still keeping the didSomeReading
guard and call select only if no file descriptors have data
available.

Note that we still go through all the connections when reading, even
though most of them might not have data. We could improve that, but
given the way select reports 'active' file descriptors (leaves their
bits set in the fd_set) we would have to go over all the clients and
check the file descriptor with FD_ISSET(fd, &fds) anyway.  (We would
however save needless read system calls, so it might be worth the
effort).

Try strace-ing this version of program, it should be easier to follow
now.